Update to SoccerDemoWithReset.py to correct inconsistent use of tabs and spaces in indentation
This may not be obvious in the github preview. Using arrow keys to check the number of characters from start may help to identify the issue.
Line 70 uses 8 spaces for indenting
Line 71 uses a Tab for indenting
Line 72 uses a Tab and 4 spaces for indenting.
Rest of the code uses only space for indenting.
The tabs in the line 71 and 72 are replaced with 4 spaces to keep it consistent throughout the code. This adheres to the PEP 8 style guide (https://peps.python.org/pep-0008/#indentation)